Feature/feature home sekar

+++ app/src/main/java/com/allana/food_recipe_app_chapter7/ui/features/favorite/FavoriteRecipeFragment.kt
@@ -8,7 +8,9 @@ import com.allana.food_recipe_app_chapter7.base.arch.BaseFragment
 import com.allana.food_recipe_app_chapter7.base.model.Resource
 import com.allana.food_recipe_app_chapter7.data.local.room.entity.FavoriteRecipe
 import com.allana.food_recipe_app_chapter7.databinding.FragmentFavoriteRecipeBinding
-import com.allana.food_recipe_app_chapter7.ui.features.favoriterecipe.adapter.FavoriteRecipeAdapter
+import com.allana.food_recipe_app_chapter7.ui.features.favorite.FavoriteRecipeContract
+import com.allana.food_recipe_app_chapter7.ui.features.favorite.FavoriteRecipeViewModel
+import com.allana.food_recipe_app_chapter7.ui.features.favorite.adapter.FavoriteRecipeAdapter
 import com.allana.food_recipe_app_chapter7.ui.features.home.detail.DetailActivity
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -32,7 +34,7 @@ FavoriteRecipeContract.View {
 
     override fun initList() {
         adapter = FavoriteRecipeAdapter {
-            DetailActivity.startActivity(context, it.idRecipe?.toInt() ?: 0)
+            DetailActivity.startActivity(context, it.idRecipe?.toLong() ?: 0)
         }
         getViewBinding().rvFavoriteRecipe.apply {
             layoutManager = LinearLayoutManager(context)
